Refactor Flight model
Replacing the relation of Flight to FlightResult with a unidirectional one. There is currently no need for a bidirectional option. Also removed is a field that can be calculated separately without having to store it in the database. And some types are being replaced.

+++ src/main/java/tech/filatov/pigeoner/repository/flight/FlightResultRepositoryCustom.java
@@ -3,5 +3,5 @@ package tech.filatov.pigeoner.repository.flight;
 import java.util.Map;
 
 public interface FlightResultRepositoryCustom {
-    Map<Long, Long> getNumberOfMyParticipantsForEveryFlight(long userId);
+    Map<Long, Integer> getNumberOfMyParticipantsForEveryFlight(long userId);
 }
